refactor: 우승자 찾기 관련 메소드 수정
- RacingTest 클래스 삭제
