+++ src/main/java/racingcar/Main.java
@@ -1,9 +1,8 @@
 package racingcar;
 
 public class Main {
-	private static final Racing racing = new Racing();
-
 	public static void main(String[] args) {
+		Racing racing = new Racing();
 		racing.start();
 	}
 }
